update "up" and "down" arrow key handlers
Since emoji button widths can vary by the count of unicode characters,
advancing either forward or backward by a constant count of emojis doesn't work.

Now, the determination of the emoji that is above or below by one row
is done based on some calculation using position of each emoji within the DOM.

Note: the plugin now takes a new option parameter: "maxResults".
This is the maximum count of emojis to display that match partial user input.
The default is: 120.